+++ Models/Transactions.cs
@@ -16,7 +16,7 @@ namespace PersonalFinanceFrontEnd.Models
         public string TrsCode { get; set; }
         public string TrsTitle { get; set; }
         public DateTime TrsDateTime { get; set; }
-        public int TrsValue { get; set; }
+        public float TrsValue { get; set; }
         public string TrsNote { get; set; }
 
     }
